Change PhaseBanner to be functional component in separate directory
